+++ inject/inject-server/src/test/scala/com/twitter/inject/server/FeatureTestMixin.scala
@@ -27,11 +27,16 @@ trait FeatureTestMixin extends SuiteMixin with IntegrationTestMixin { this: Suit
 
   override protected def afterEach(): Unit = {
     super.afterEach()
-    if (server.isInjectable) {
-      if (printStats) {
-        server.printStats()
+    try {
+      if (server.usesInMemoryStatsReceiver) {
+        if (printStats) {
+          server.printStats()
+        }
+        server.clearStats()
       }
-      server.clearStats()
+    } catch {
+      case _: IllegalStateException => /* DO NOTHING */
+      // we don't have access to a StatsReceiver to perform these functions for the user
     }
   }