+++ inject/inject-server/src/test/scala/com/twitter/inject/server/tests/FeatureTestNonInjectionCustomStatsReceiverTest.scala
@@ -3,15 +3,18 @@ package com.twitter.inject.server.tests
 import com.twitter.finagle.http.Status
 import com.twitter.inject.server.{EmbeddedTwitterServer, FeatureTest}
 
-/** Test a non-inject TwitterServer with the [[FeatureTest]] trait */
-class FeatureTestNonInjectionTest extends FeatureTest {
+/** Test a non-inject TwitterServer using a Custom StatsReceiver implementation with the [[FeatureTest]] trait */
+class FeatureTestNonInjectionCustomStatsReceiverTest extends FeatureTest {
+
+  private[this] val testStatsReceiver: TestStatsReceiver = new TestStatsReceiver
 
   override val server: EmbeddedTwitterServer =
     new EmbeddedTwitterServer(
-      twitterServer = new NonInjectionTestServer(),
+      twitterServer = new NonInjectionTestServer(Some(testStatsReceiver)),
       args = Seq("http.port=:0"),
-      disableTestLogging = true)
-
+      disableTestLogging = true,
+      statsReceiverOverride = Some(testStatsReceiver)
+    )
   /* Does not work since this we do not provide an InMemoryStatsReceiver but should not result in failures */
   override val printStats: Boolean = true
 
@@ -28,15 +31,15 @@ class FeatureTestNonInjectionTest extends FeatureTest {
   }
 
   test("TwitterServer#stats receivers") {
-    // for a non-injectable server that has provided no SR override
-    // we have no way of providing anything useful here
-    intercept[IllegalStateException] {
-      server.statsReceiver
-    }
-    // same as above, which means we also can't provide an in memory stats receiver
+    // even though the server under test is a non-injectable server, we
+    // have been provided with a stats receiver override which we return here
+    server.statsReceiver
+    // but it is not in-memory stats receiver
     intercept[IllegalStateException] {
       server.inMemoryStatsReceiver
     }
+
+    assert(testStatsReceiver.gauges.nonEmpty) /* we add a build revision gauge in startup of the server */
   }
 
   test("TestServer#feature test") {
@@ -50,5 +53,4 @@ class FeatureTestNonInjectionTest extends FeatureTest {
       andExpect = Status.Ok
     )
   }
-
 }